De-duplicated matrix generators, moved logging to logging dir instead of utils
