--- examples/arnoldi/generators/value_factory.hh
@@ -1,9 +1,0 @@
-#pragma once
-
-#include <random>
-
-template<typename T>
-class value_factory {
-public:
-    virtual T next() = 0;
-};
